+++ shippo/config.py
@@ -1,9 +1,12 @@
+import os
+
 # Configurable variables
-api_key = None
-api_base = 'https://api.goshippo.com/'
-api_version = '2018-02-08'
+api_key = os.environ.get('SHIPPO_API_KEY')
+api_base = os.environ.get('SHIPPO_API_BASE', 'https://api.goshippo.com/')
+api_version = os.environ.get('SHIPPO_API_VERSION', '2018-02-08')
 verify_ssl_certs = True
 rates_req_timeout = 20.0
-
-# None falls back to 80s for RequestsClient and 55s for UrlFetchClient
 timeout_in_seconds = None
+app_name = 'PythonApp'
+app_version = '1.0'
+vcr_logging_level = 'ERROR'