+++ shippo/resource.py
@@ -2,7 +2,6 @@ import urllib.parse
 import sys
 import time
 import warnings
-import shippo.config as config
 from shippo import api_requestor, error, util
 from shippo.config import rates_req_timeout
 
@@ -165,7 +164,7 @@ class APIResource(ShippoObject):
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "/v1/%ss" % (cls_name,)
+        return "%ss" % (cls_name,)
 
     def instance_url(self):
         object_id = self.get('object_id')
@@ -256,21 +255,32 @@ class Address(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ses/" % (cls_name,)
+        return "%ses/" % (cls_name,)
 
 
 class CustomsItem(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
 
     @classmethod
     def class_url(cls):
-        return "v1/customs/items/"
+        return "customs/items/"
 
 
 class CustomsDeclaration(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
 
     @classmethod
     def class_url(cls):
-        return "v1/customs/declarations/"
+        return "customs/declarations/"
+
+
+class Order(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
+    """
+    An Order allows you group line items for a shipment.
+    """
+
+    @classmethod
+    def class_url(cls):
+        cls_name = cls.class_name()
+        return "%ss/" % (cls_name,)
 
 
 class Parcel(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
@@ -278,7 +288,18 @@ class Parcel(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
+
+
+class Pickup(CreateableAPIResource):
+    """
+    A Pickup allows you to schedule pickups with USPS and DHL Express for eligible shipments that you have already created.
+    """
+
+    @classmethod
+    def class_url(cls):
+        cls_name = cls.class_name()
+        return "%ss/" % (cls_name,)
 
 
 class Manifest(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
@@ -290,7 +311,7 @@ class Manifest(CreateableAPIResource, ListableAPIResource, FetchableAPIResource)
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
 
 
 class Refund(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
@@ -301,7 +322,7 @@ class Refund(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
 
 
 class Shipment(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
@@ -331,7 +352,7 @@ class Shipment(CreateableAPIResource, ListableAPIResource, FetchableAPIResource)
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
 
 
 class Transaction(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
@@ -352,7 +373,7 @@ class Transaction(CreateableAPIResource, ListableAPIResource, FetchableAPIResour
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
 
 
 class Rate(ListableAPIResource, FetchableAPIResource):
@@ -364,14 +385,15 @@ class Rate(ListableAPIResource, FetchableAPIResource):
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
 
 
 class CarrierAccount(CreateableAPIResource, ListableAPIResource, FetchableAPIResource, UpdateableAPIResource):
 
     @classmethod
     def class_url(cls):
-        return "v1/carrier_accounts/"
+        return "carrier_accounts/"
+
 
 class Webhook(CreateableAPIResource, ListableAPIResource, FetchableAPIResource, UpdateableAPIResource):
     """
@@ -383,7 +405,7 @@ class Webhook(CreateableAPIResource, ListableAPIResource, FetchableAPIResource, 
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
 
     @classmethod
     def list_webhooks(cls, api_key=None, **params):
@@ -431,7 +453,6 @@ class Webhook(CreateableAPIResource, ListableAPIResource, FetchableAPIResource, 
             (ShippoObject) -- The server response
         """
         return super(Webhook, cls).remove(object_id,api_key, **params)
-
 
 
 class Track(CreateableAPIResource):
@@ -493,7 +514,7 @@ class Track(CreateableAPIResource):
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ss/" % (cls_name,)
+        return "%ss/" % (cls_name,)
 
 
 class Batch(CreateableAPIResource, FetchableAPIResource):
@@ -607,4 +628,26 @@ class Batch(CreateableAPIResource, FetchableAPIResource):
     @classmethod
     def class_url(cls):
         cls_name = cls.class_name()
-        return "v1/%ses/" % (cls_name,)
+        return "%ses/" % (cls_name,)
+
+
+class Order(CreateableAPIResource, ListableAPIResource, FetchableAPIResource):
+    """
+    Beta functionality, but useful for tracking a given order, label information, etc.
+    """
+
+    @classmethod
+    def class_url(cls):
+        cls_name = cls.class_name()
+        return "%ss/" % (cls_name,)
+
+
+class LineItem(ListableAPIResource, FetchableAPIResource):
+    """
+    Currently this is a nested item inside an Order object.
+    """
+
+    @classmethod
+    def class_url(cls):
+        cls_name = cls.class_name()
+        return "%ss/" % (cls_name,)